Slightly improve the setting page tables
